Fix #6: Page administration broken with new page splitting
Basically, we have to change the regexps to deal with the split
comments instead of the headings as before. We can simplify the regexps
because the page title in the new split comments doesn't contain markup
anymore. We also have to insert the split marker for newly created
pages, instead of the old heading marker.

Instead of `$cf[menu][levels]` we have to use the maximum nesting
level, which is currently 9 (a constant appears to be useful here, by
the way).