Close #1: Upgrade to jsTree 3
We finally manage to upgrade to jsTree 3.3.3. From the user perspective
the pagemanager widget almost works as before, with two minor changes:
unrenameable and duplicate pages are marked with an icon. The former
were not marked earlier at all, and the latter were marked by changing
the page title to "DUPLICATE HEADING X". Doing this with jsTree 3 is
way too much trouble, and actually one may argue that showing the
actual page title is an improvement.