Fix #36: Pages marked as unrenameable can be renamed
We make sure that unrenameable pages can't be renamed in the widget.
Unfortunately it doesn't appear to be possible to actually prevent the
edit mode, but at least we can reject any changes after user editing.
As it's unlikely that pages names are unrenameable under XH 1.7,
because it doesn't make sense anymore to add HTML tags to the page
names, this solution seems to be sufficient.