Close #48: Default theme is unsuitable for dark templates
We are reluctant to change the default theme (or to override some theme
rules in Pagemanager's stylesheet), so we add the "proton" theme,
slightly adapt it to our needs, and make it the default. A particular
advantage of the "proton" theme are its checkbox images which fit much
better to the Font Awesome icons then the colored ones of the default
theme.