error with use reserved characters
Many problems because reserved characters are not excluded or replaced.
In case of an update it does not come into effect, so it does not cause any problems and in case of a new installation it would be possible.

https://github.com/cmsimple-xh/cmsimple-xh/issues/413

If all reserved characters are replaced with $tx['urichar']['org'], $tx['urichar']['new'], then under certain situations the page manager will delete the complete content of the content.html.

https://github.com/cmb69/pagemanager_xh/issues/63