Fix #66: Pagemanager_XH creates erroneous preview/edit links
Obviously, the `Url` abstraction of Pagemanager_XH is insufficient for
CMSimple_XH, since the first parameter of the query string may have
special meaning, and can usually be re-used from `$su` or `$u`, where
it is already properly `uenc()`'d.  However, because the URL handling
of CMSimple_XH appears to be in need of a general clean-up, and since
we don't a need a general handling for Pagemanager_XH right now, we
work-around that issue by simply constructing these URLs manually.