Switch dev tools from composer to phive where possible
This mitigates the dependency hell, and also lets us quickly change
PHPUnit versions for testing with different PHP versions.

The downside is that it introduces yet another tool, but we hope to get
rid of the composer dev requirements in the long run.

To be clear, composer is a *great* tool, but not so for dev
requirements.