Fix `coverage` target wrt. phpunit.phar
We switched to phpdbg as coverage provider to improve the performance,
since Xdebug 2 is very slow.  However, the performance of Xdebug 3
greatly improved, and there is also pcov which may even be faster.  The
latter requires PHPUnit 8 (or pcov-clobber), but for the few tests,
using Xdebug 2 is not really a problem.