Fix #71: Prevent saving linked_to_menu as published and vice versa
We store the current pagedata_attribute in a hidden form field, and
reject form submission if its value has changed.  That should be good
enough to cater to this pathological case.

We also drop the PHPStan baseline, and replace it with a more general
error suppression regarding unknown View attributes.